+++ nes-nautilus/include/Nautilus/Interface/BufferRef/RowTupleBufferRef.hpp
@@ -20,6 +20,8 @@
 #include <DataTypes/DataType.hpp>
 #include <Nautilus/Interface/BufferRef/TupleBufferRef.hpp>
 #include <Nautilus/Interface/Record.hpp>
+#include <Runtime/AbstractBufferProvider.hpp>
+#include <val.hpp>
 
 namespace NES
 {
@@ -62,13 +64,13 @@ public:
         const RecordBuffer& recordBuffer,
         nautilus::val<uint64_t>& recordIndex) const override;
 
-    nautilus::val<size_t> writeRecord(
+    nautilus::val<uint64_t> writeRecord(
         nautilus::val<uint64_t>& recordIndex,
         const RecordBuffer& recordBuffer,
         const Record& rec,
         const nautilus::val<AbstractBufferProvider*>& bufferProvider) const override;
 
-    nautilus::val<size_t> writeRecordSafely(
+    nautilus::val<uint64_t> writeRecordSafely(
         nautilus::val<uint64_t>& recordIndex,
         const RecordBuffer& recordBuffer,
         const Record& rec,
